Implemented the obstetricsInit DB
The JSF backend setup for the obstetrics initialization page has been
setup.  This includes the DB and test values to populate the DB.
Currently, Patient 1 is the only obstetrics eligible patient with
records in the DB.  The MySQL, SQLLoader, Validator, and POJO object
have all been setup according to iTrust specifications.