Add UI elements to the Obstetrics tab in Office Visit.
UI Changes made:
 - Tabs were added to the Office Visit page for Obstetrics and
Ultrasound.
 - The Obstetrics page has the necessary UI elements present, except for
the save/cancel buttons.
 - The xhtml file for the Ultrasound tab exists, but does not contain
any UI elements.
 - None of the newly added UI elements are currently tied to any
functional code.

New Controller and Form:
 - ObstetricsVisitForm is partially implemented. Most of its real code
is currently replaced with dummy code due to the lack of a functional
backend.
 - ObstetricsVisitController is partially implemented. Most of its real
code is currently replaced with dummy code due to the lack of a
functional backend. However,
ObstetricsVisitController.calculateWeeksPregnant(OfficeVisit) should be
functional (but it has not been tested).

Changes made to test data and UC93 Unit Tests:
 - The currently existing test data has been shifted around so that the
obstetrics initialization entries are in order of date (which is what
would be expected in real use).
 - To account for the changes in test data, the following two test
classes required small adjustments:
    - ObstetricsInitMySQLTest.java
    - PregnancyInfoMySQLTest.java