One more Unit Test
This unit tests has two conflicts two days in a row.  The first is from
a an existing appointment, where the start time of the new appointment
conflicts with it.  The second is from another appointment created from
a previous unit test, where the end time of the new appointment
conflicts with it.