Fix failing cucumber tests not related to childbirth visit tab.
Additionally, change the global messages to display as black if faces
message severity info is used and display red if severity error is used.

To test equality of some Newborn objects, the checking of matching PID
field had to be removed. This is because there was no good way for me to
find the generated PID and update my Newborn instances.