Fix code smell 1.
Add method addLogOfficeVisitID() to class iTrustController. This class
is called by PrescriptionController.add() and
MedicalProcedureController.add(), which now delegate to
iTrustController.addLogOfficeVisitID().

This fix required casting PrescriptionMySQL and MedicalProcedureMySQL as
a DataBean. Unfortunately, they were not implementing that interface
(although they should have been to match the architectural design
pattern used throughout iTrust), so this required several changes to
source and test code that are not directly related to the refactoring
itself.